+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - IssuingStockRule resource
 - Rules and DebtorWorkspaceID attributes to CreditNote resource
+- PixFraud resource
 
 ## [0.12.0] - 2026-02-10
 ### Fixed